retry old working files before I restarted the App

+++ back-end/src/db/migrations/20220307092900_createTablesTable.js
@@ -1,18 +1,21 @@
 exports.up = function (knex) {
-  return knex.schema.createTable("tables", (table) => {
-    table.increments("table_id").primary();
-    table.string("table_name");
-    table.integer("capacity");
-    table.string("status").defaultTo("free");
-    table
-      .integer("reservation_id")
-      .references("reservation_id")
-      .inTable("reservations")
-      .onDelete("set null");
-    table.timestamps(true, true);
+  return knex.schema.hasTable('tables').then(function(exists) {
+    if (!exists) {
+      return knex.schema.createTable('tables', function(table) {
+        table.increments('table_id').primary();
+        table.string('table_name');
+        table.integer('capacity');
+        table.string('status').defaultTo('free');
+        table.integer('reservation_id')
+             .references('reservation_id')
+             .inTable('reservations')
+             .onDelete('set null');
+        table.timestamps(true, true);
+      });
+    }
   });
 };
 
 exports.down = function (knex) {
-  return knex.schema.dropTable("tables");
+  return knex.schema.dropTable('tables');
 };
